pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "helsinki-localstack-pipeline"
version = "0.1.0"
description = "Local AWS data pipeline with LocalStack, Airflow, Spark, Lambda, and DynamoDB"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "boto3==1.34.162",
  "pandas==2.2.3",
  "python-dotenv==1.0.1",
  "pyspark==3.5.2"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
